+++ new_backend/StationApp/Dtos/StationTemperatureReadDto.cs
@@ -10,7 +10,7 @@ namespace StationApp.Dtos
         [Key]
         public int TemperatureId { set; get; }
         
-        public int Temperature{ set; get; }
+        public float Temperature{ set; get; }
 
         public long Time { get; set; }
